requirements.txt
Django==4.1.1
django-tinymce==3.5.0
python-dotenv==0.21.0

